version 0.6 - january 2010
1. Modified validate tool to check for unsorted vcf.  This checks
to ensure that the records are sorted numerically within a reference
sequence and that once a reference sequence has been seen, it
cannot be seen again (unless records appear consecutively).  For example,
if 1:4, 1:7, 2:4 is fine, but if another record for reference sequence 1
was to appear, the vcf file would be flagged as unsorted.  The union
and intersection tools may not function correctly on unsorted vcf files
(it is fine if the reference sequences are ordered differently, however).

2. Completed the union tool.  Written to ensure that nothing is stored in
memory.

3. Intersection tool modified.