version 0.21 - february 2011
1. If input files have no header, a couple of header lines are
generated and fed to the output.  Previously, no header was
generated which caused a problem when piping as the stdin would
have no header and the header parser would try to close and reopen
the input to ensure that a record wasn't missed.  This would cause
an error as the stdin could not be closed and reopened.