version 0.22 - february 2011
1. The intersect and union tools now allow the --priority-file (-f)
argument to be set to "merge".  In this case, the information from
both vcf files will be propogated into the new file seperated by a "/".
The header gives a list of all of the files used to generate the list
of data and this order is preserved in all records.